inference/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aioice==0.9.0
aiortc==1.8.0
aiosignal==1.3.1
altair==5.3.0
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
av==11.0.0
bitarray==2.8.2
bitstring==4.1.2
blinker==1.8.1
build==0.10.0
CacheControl==0.13.1
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cleo==2.0.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
crashtest==0.4.1
cryptography==42.0.6
cycler==0.12.1
distlib==0.3.7
Django==4.1.5
dnspython==2.6.1
dulwich==0.21.6
filelock==3.12.4
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.5.0
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.8.0
importlib_resources==6.4.0
imutils==0.5.4
installer==0.7.0
jaraco.classes==3.3.0
jax==0.4.26
jaxlib==0.4.26
Jinja2==3.1.3
jsonschema==4.17.3
keyring==24.2.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mediapipe==0.10.13
ml-dtypes==0.4.0
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.5
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.2
pexpect==4.8.0
pillow==10.3.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
protobuf==4.25.3
ptyprocess==0.7.0
pyarrow==16.0.0
pycparser==2.22
pydeck==0.9.0
pyee==11.1.0
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
rapidfuzz==2.15.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
sqlparse==0.4.3
streamlit==1.34.0
streamlit-webrtc==0.47.1
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.1
tornado==6.4
trove-classifiers==2023.10.18
typing_extensions==4.11.0
tzdata==2022.7
urllib3==2.0.7
virtualenv==20.24.6
watchdog==4.0.0
webcam==1.34
yarl==1.9.4
zipp==3.17.0
